oslayer_iomgr_test: widen rand() before *4096 to avoid signed-int UB
InitIoreq builds a fake pbData pointer as `rand() * 4096 | 0xC000420000000000`
to plant the IOREQ buffer in a wild upper-half address.  On Windows MSVC
RAND_MAX is 0x7FFF so the product fits in a 32-bit int.  On Linux glibc
RAND_MAX is 0x7FFFFFFF and `rand() * 4096` overflows signed int — UB that
clang -O3 happily folds into a negative int, which sign-extends and ORs
into pointers like 0xffffffffffffe000.

TestReadCombiningProducesSgioPreparableIorunCombinatorial walks the IO
chain with `ulpLow + cbData < ulpHigh` checks; when pbData lives in the
last page of the address space, that 64-bit unsigned addition wraps to a
tiny value and the range check rejects every FSE entry.  Stack garbage
from the preceding non-combinatorial test happened to mask the bug for
the wildcard-pattern run but not when the combinatorial test ran first.

Cast rand() to ULONGLONG so the multiply happens at 64-bit width.

+++ dev/ese/src/ese/oslayer_iomgr_test.cxx
@@ -105,7 +105,12 @@ void InitIoreq( IOREQ * const pioreq, _OSFILE& _osf, const BOOL fWrite, const QW
     //  The pointer doesn't matter as we don't deref it or do IO with it, but we do test it for non-null
     //  and alignment expectations.  Randomly assigning it to the address space will (probably) AV when
     //  it is deref'd, so that's nice.
-    pioreq->pbData = (BYTE*)( rand () * 4096
+    //
+    //  Cast rand() to ULONGLONG before multiplying.  On Windows RAND_MAX is 0x7FFF so rand()*4096 fits
+    //  in 32 bits, but on Linux glibc RAND_MAX is 0x7FFFFFFF and rand()*4096 overflows signed-int —
+    //  clang -O3 takes the UB and yields a negative int that sign-extends to a top-of-address-space
+    //  pointer (e.g. 0xffffffffffffe000), causing FCheckOffsetInChainOrGap's `ulpLow + cbData` to wrap.
+    pioreq->pbData = (BYTE*)( (ULONGLONG)rand () * 4096
 #ifdef ESE_ARCH_64BIT
                         | 0xC000420000000000
 #else
